Enhancements on edit user
