+++ frontend/SW-Quality-Dashboard/src/components/ui/DashboardNavbar/DashboardNavbar.tsx
@@ -11,7 +11,11 @@ function DashboardNavbar({isAuthenticated, isAdmin}) {
 
     return (
         <>
-            <Navbar expand="lg" className="bg-body-tertiary" style={{borderBottom: "2px solid var(--org-color-primary-dark-version)", height: "10vh"}}>
+            <Navbar expand="lg" className="bg-body-tertiary" style={{
+                borderBottom: "2px solid var(--org-color-primary-dark-version)",
+                minHeight: "10vh",
+                padding: "0 2rem"
+            }}>
                 <Container>
                     <Navbar.Brand href="/" className="navbar-brand">
                         <img src={logo}/>
