--- frontend/SW-Quality-Dashboard/src/pages/admin/AdminAllUsersPage/AdminAllUsersPage.css
@@ -1,2 +1,0 @@
-
-
