+++ frontend/SW-Quality-Dashboard/src/pages/admin/AdminAllUsersPage/AdminAllUsersPage.tsx
@@ -8,6 +8,8 @@ import axios from "axios";
 import {Alert, Badge, Button, Form, Modal, Table} from "react-bootstrap";
 import React from "react";
 import AllUsersTable from "../tables/AllUsersTable/AllUsersTable.tsx";
+import {Heading} from "@chakra-ui/react";
+import EditUserModal from "../../../components/modals/EditUserModal/EditUserModal.tsx";
 
 const baseUrl = import.meta.env.VITE_API_BASE_URL;
 
@@ -55,55 +57,6 @@ function AdminAllUsersPage() {
             })
     }, [accessToken]);
 
-    const handleEditSubmit = () => {
-        console.log('Edit User')
-        // Check if there are any changes
-        let noChanges = editedName === editUser.name && editedEmail === editUser.email;
-        let emptyInput = editedName === '' && editedEmail === '';
-        if (noChanges || emptyInput) {
-            setEditAlertVariant('warning')
-            setEditAlertMessage('No changes were made')
-            setEditAlert(true)
-            return
-        }
-
-        // Make a map only with the provided values
-        let editedUser = {}
-        if (editedName !== '') {
-            editedUser['name'] = editedName
-        }
-        if (editedEmail !== '') {
-            editedUser['email'] = editedEmail
-        }
-
-        // Call the api to update the user
-        let url = baseUrl + apiUrls.routes.admin.updateUser;
-        url = url.replace(':userId', editUser.id);
-
-        let headers = {
-            'Authorization': 'Bearer ' + accessToken,
-            'Content-Type': 'application/json'
-        }
-
-        axios.put(url, editedUser, { headers: headers })
-            .then(response => {
-                setEditAlertVariant('success')
-                setEditAlertMessage('User updated successfully')
-                setEditAlert(true)
-
-                // Wait 5 seconds before closing the modal
-                setTimeout(() => {
-                    setShowEditUserModal(false)
-                }, 5000)
-            })
-            .catch(error => {
-                console.error(error)
-                setEditAlertVariant('danger')
-                setEditAlertMessage('Error updating user')
-                setEditAlert(true)
-            })
-    }
-
     return (
         <AdminTabContent
             icon="bi bi-people-fill"
@@ -119,54 +72,12 @@ function AdminAllUsersPage() {
                 The user with <strong>bold</strong> is you.
             </p>
 
-            {/* Edit User Modal */}
-            <Modal
+            <EditUserModal
                 show={showEditUserModal}
-                onHide={() => setShowEditUserModal(false)}
-                backdrop="static"
-                keyboard={false}
-            >
-                <Modal.Header closeButton>
-                    <Modal.Title>
-                        <i className="bi bi-person-fill-gear"> </i>
-                        Edit User
-                    </Modal.Title>
-                </Modal.Header>
-                <Modal.Body>
-                    {showEditUserModal && (
-                        <Alert variant={editAlertVariant}>
-                            <i className={editAlertVariant === 'danger' || editAlertVariant === 'warning' ? 'bi bi-exclamation-triangle-fill' : 'bi bi-check-circle-fill'}> </i>
-                            {editAlertMessage}
-                        </Alert>
-                    )}
-                    <Form>
-                        <Form.Group className="mb-3" controlId="formBasicEmail">
-                            <Form.Label>
-                                <i className="bi bi-person-fill"> </i>
-                                Username
-                            </Form.Label>
-                            <Form.Control type='text' placeholder={editUser.name} onChange={(e) => setEditedName(e.target.value)} />
-                        </Form.Group>
-
-                        <Form.Group className="mb-3" controlId="formBasicPassword">
-                            <Form.Label>
-                                <i className="bi bi-envelope-fill"> </i>
-                                Email
-                            </Form.Label>
-                            <Form.Control type="email" placeholder={editUser.email} onChange={(e) => setEditedEmail(e.target.value)} />
-                        </Form.Group>
-                    </Form>
-
-                </Modal.Body>
-                <Modal.Footer>
-                    <Button variant="secondary" onClick={() => setShowEditUserModal(false)}>
-                        Cancel
-                    </Button>
-                    <Button variant="primary" onClick={handleEditSubmit}>
-                        Save
-                    </Button>
-                </Modal.Footer>
-            </Modal>
+                setShow={setShowEditUserModal}
+                handleClose={() => setShowEditUserModal(false)}
+                editUser={editUser}
+            />
 
             {/* Delete User Modal */}
             <Modal
